Remove risk line
>I am comfortable diving head-first into moderately risky environments with relatively little upfront information and oversight.

This, to me suggests that we are not planning our work with clients appropriately before sending out consultants. I suggest we remove it and also establish a working practise that doesn't allow this situation (accepting edge cases). If a team is to be sent out to a client then they should be involved much earlier in any delivery processes (scoping, design etc) to ensure that everyone is is up to speed.